+++ src/components/search-bar-operator/SearchBarOperator.stories.jsx
