+++ src/components/search-bar-operator/SearchBarOperator.tsx
