+++ src/components/search-bar-operator/index.ts
