+++ src/components/search-bar-operator/styles/SSearchBarOperator.tsx
