+++ src/components/search-bar-operator/styles/SSearchBarOperatorWrapper.tsx
